requirements.txt
# Obsidian PKM System Requirements

# YAML parsing for frontmatter
PyYAML>=6.0.1

# Environment variables (.env file support)
python-dotenv>=1.0.0

# Google Gemini API for Atomic Note Agent (Stage 1)
google-generativeai>=0.8.0

# NER (Named Entity Recognition) - Stage 2
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0

# Future dependencies
# neo4j>=5.14.0  # For Graph DB (Stage 3)

